Fix Y-shearing sprite math overflow and render full 3D view around centered status bar

+++ files/r_segs.c
@@ -174,7 +174,7 @@ R_RenderMaskedSegRange
 		dc_colormap = walllights[index];
 	    }
 			
-	    sprtopscreen = centeryfrac - FixedMul(dc_texturemid, spryscale);
+	    sprtopscreen = (int64_t)centeryfrac - (int64_t)FixedMul(dc_texturemid, spryscale);
 	    dc_iscale = 0xffffffffu / (unsigned)spryscale;
 	    
 	    // draw the texture
